+++ src/main/kotlin/practicaenequipocristianvictoraitornico/personal/dto/EntrenadorDto.kt
@@ -1,0 +1,19 @@
+package practicaenequipocristianvictoraitornico.personal.dto
+
+
+/**
+ * Clase que representa a los entrenadores.
+ *
+ * @param especialidad Tipo de entrenador.
+ */
+
+class EntrenadorDto(
+    id: Long,
+    nombre: String,
+    apellidos: String,
+    fechaNacimiento: String,
+    fechaIncorporacion: String,
+    salario: Double,
+    pais: String,
+    val especialidad: String
+): PersonaDto(id, nombre, apellidos, fechaNacimiento, fechaIncorporacion, salario, pais)
