+++ src/main/kotlin/practicaenequipocristianvictoraitornico/personal/dto/EquipoDtoXml.kt
@@ -1,0 +1,12 @@
+package practicaenequipocristianvictoraitornico.personal.dto
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+
+/**
+ * DTO que representa la estructura del XML de equipo.
+ */
+@Serializable
+@SerialName("equipo")
+data class EquipoDtoXml(val personal: List<PersonalDtoXml>)
