+++ src/main/kotlin/practicaenequipocristianvictoraitornico/personal/dto/JugadorDto.kt
@@ -1,0 +1,18 @@
+package practicaenequipocristianvictoraitornico.personal.dto
+
+
+class JugadorDto(
+    id: Long,
+    nombre: String,
+    apellidos: String,
+    fechaNacimiento: String,
+    fechaIncorporacion: String,
+    salario: Double,
+    pais: String,
+    val posicion: String,
+    val dorsal: Int,
+    val altura: Double,
+    val peso: Double,
+    val goles: Int,
+    val partidosJugados: Int
+): PersonaDto(id, nombre, apellidos, fechaNacimiento, fechaIncorporacion, salario, pais)
