+++ src/main/kotlin/practicaenequipocristianvictoraitornico/personal/validator/Validate.kt
@@ -1,0 +1,7 @@
+package practicaenequipocristianvictoraitornico.personal.validator
+
+import com.github.michaelbull.result.Result
+
+interface Validate<T,E> {
+    fun validator(item: T): Result<T,E>
+}
