--- src/main/kotlin/practicaenequipocristianvictoraitornico/players/validator/Validate.kt
@@ -1,7 +1,0 @@
-package practicaenequipocristianvictoraitornico.players.validator
-
-import com.github.michaelbull.result.Result
-
-interface Validate<T,E> {
-    fun validator(item: T): Result<T,E>
-}
