--- src/main/kotlin/test/player/storage/PersonalStorageJsonTest.kt
@@ -1,5 +1,0 @@
-package test.player.storage
-
-class PersonalStorageJsonTest {
-
-}
