+++ src/test/kotlin/player/storage/PersonalStorageBinTest.kt
@@ -1,0 +1,4 @@
+package player.storage
+
+class PersonalStorageBinTest {
+}
