+++ src/test/kotlin/player/storage/PersonalStorageCsvTest.kt
@@ -1,0 +1,4 @@
+package player.storage
+
+class PersonalStorageCsvTest {
+}
