+++ src/test/kotlin/player/storage/PersonalStorageJsonTest.kt
@@ -1,0 +1,5 @@
+package player.storage
+
+class PersonalStorageJsonTest {
+
+}
