+++ src/test/kotlin/player/storage/PersonalStorageXmlTest.kt
@@ -1,0 +1,4 @@
+package player.storage
+
+class PersonalStorageXmlTest {
+}
